+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Update Multichain to v0.3.21
 - Update Solana recipient encoding
 - Add support for Goerli
+- Fix gateway querying for Filecoin
 
 ## 0.4.5